moving batch counting and migration verification
